requirements.txt
djangorestframework
mysqlclient
django-cors-headers
